% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptHomemadeFunctions.R
\name{IC95}
\alias{IC95}
\title{Ninety-five percent confidence interval}
\usage{
IC95(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The range of a confidence interval.
}
\description{
Computes the 95 percent confidence interval of a mean. Only trustworthy for normally distributed data (or data with n > 30).
}
\examples{
myvector <- rnorm(n = 100, mean = 48, sd = 16)
IC95(x = myvector)
# Gives a result of 2.992398. It means that the mean of myvector
# has 95\% chance to be comprised in an interval of +/- 2.992398.
}
