% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptHomemadeFunctions.R
\name{posthoc.NMDS}
\alias{posthoc.NMDS}
\title{Two-by-two post-hoc tests for NMDSs using PERMANOVA}
\usage{
posthoc.NMDS(REL, FAC)
}
\arguments{
\item{REL}{A contingency table. Typically, a species-site matrix.}

\item{FAC}{A factor containing the different modalities/treatments we want to combine and test.}
}
\value{
A table containing the p-values and corrected p-values for each PERMANOVA test.
}
\description{
This function performs as many NMDSs and PERMANOVAs as there are possible combinations between the input factor levels. In other words, the
function will compute (using the \code{\link[airpoumpoum:comb_mod]{comb_mod}} function) all possible combinations of the modalities (levels) contained in
\emph{FAC}, and run a separate NMDS and associated PERMANOVA between each pair of those modalities (i.e. treatments) and the community data in \emph{REL}. \cr
The post-hoc tests are corrected using \strong{Holm-Bonferroni} method to control the family-wise error rate. By default, the functions only performs
NMDSs and PERMANOVAs using \strong{Bray-Curtis} distances, and keep only 2 dimensions in each ordination.
}
\note{
This function should only be used when a preliminary PERMANOVA detected a significant effect of \emph{FAC} on the centroid position of \emph{REL},
and after ensuring that there was no \emph{multivariate spread} in the data groups (see \code{\link[vegan:adonis]{adonis}}).
}
\examples{
## For a simulated matrix with 20 lines (sites) and
## 30 columns (species), split in 2 treatments:
library(vegan)
data(dune)
mytreatment <- gl(2, 10)
aa <- posthoc.NMDS(REL = dune, FAC = mytreatment)
aa
## Here, since there is only 2 levels in the factor, there
## will only be one combination and thus, one test.
## It's a silly example allright.
}
