% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptHomemadeFunctions.R
\name{rar.rm_V2}
\alias{rar.rm_V2}
\title{Uncommon species remover}
\usage{
rar.rm_V2(tableau.AD, NB_REL = 1, ABUN = 1e-08)
}
\arguments{
\item{tableau.AD}{A contingency table (typically with species as columns and sites/releves as lines).}

\item{NB_REL}{The minimum number of sites in which the species needs to be oberved to be kept.}

\item{ABUN}{The minimum number of observations within site required for the species to be kept.}
}
\value{
A contingency table with only the species that are found at least ABUN times in NB_REL different sites.
}
\description{
Enables removing uncommon species from a contingency table (matrix or dataframe) as a function of their abundance
(number of obervations per site) and frequency (number of site where the species has been observed). Useful to only keep common and structuring
species in community data.
}
\examples{
#Nope
}
